A fix for CLion + Windows combo
This fix helps to make SFML work with CLion+Windows without messing with CLion+Linux. To make it work, you need to put a folder with SFML library (GCC 7.3.0 MinGW (SEH) - 64-bit, most likely) in a folder next to the_ark folder (starting SFML_DIR with ../ means that we move one folder up from the CMakeLists.txt). After that it compiles and links. but to execute it you also need to copy all the .dll files from SFML (bin folder) next to the .exe file.